Backend/controller/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
pytest-django==4.9.0
pytest==8.3.4
playwright==1.49.1
pytest-bdd==8.1.0
pytest-playwright==0.6.2
